Moved Controller.cs to Utils namespace
- Changed Cache size from 3 to 2 as it was meant to be
